+++ src/common/models/billingmodels/ratedusageoneof/operatingsystemrecord.go
@@ -1,4 +1,4 @@
-package billingmodels
+package ratedusageoneof
 
 import (
 	"time"
