+++ src/common/models/billingmodels/ratedusageoneof/publicsubnetdetails.go
@@ -1,4 +1,4 @@
-package billingmodels
+package ratedusageoneof
 
 import "github.com/phoenixnap/go-sdk-bmc/billingapi"
